Applied Eclipse's cleanup function on all projects with very conservative settings.
git-svn-id: https://svn.jfire.org/svn/main/trunk/org.nightlabs.base.ui@9055 18c61ebf-f106-0410-9e1e-81fb82415b8f

+++ src/org/nightlabs/base/ui/timepattern/builder/MonthlyTimePatternBuilderHopPage.java
@@ -84,7 +84,7 @@ public class MonthlyTimePatternBuilderHopPage extends WizardHopPage {
 				if (dayButton.getSelection())
 					selectedDays.add((Integer)dayButton.getData());
 				else
-					selectedDays.remove((Integer)dayButton.getData());
+					selectedDays.remove(dayButton.getData());
 			}
 		}
 		
@@ -99,7 +99,7 @@ public class MonthlyTimePatternBuilderHopPage extends WizardHopPage {
 				if (dayButton.getSelection())
 					selectedMonths.add((Integer)dayButton.getData());
 				else
-					selectedMonths.remove((Integer)dayButton.getData());
+					selectedMonths.remove(dayButton.getData());
 			}
 		}
